Lexer and Syntax into Decoder.
